Move TryGetCodeAndSpan to new MarkupCode class
